# apilaravel-main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Biblioteca",
        "description": "Documentação da API do projeto de pós-graduação",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/livros": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Lista todos os livros",
                "description": "Retorna uma lista de livros",
                "operationId": "getLivrosList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Livro"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Cria um novo livro",
                "description": "Cria um novo livro e o retorna",
                "operationId": "storeLivro",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "autor_id",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Dom Casmurro"
                                    },
                                    "autor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/livros/{id}": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Obtém um livro específico",
                "description": "Retorna os dados de um livro",
                "operationId": "getLivroById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualiza um livro existente",
                "description": "Atualiza os dados de um livro e o retorna",
                "operationId": "updateLivro",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Memórias Póstumas de Brás Cubas"
                                    },
                                    "autor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livro atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Deleta um livro",
                "description": "Deleta um livro existente",
                "operationId": "deleteLivro",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro deletado com sucesso"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Livro": {
                "title": "Livro",
                "required": [
                    "titulo",
                    "autor_id",
                    "categoria_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID do livro",
                        "type": "integer",
                        "format": "int64"
                    },
                    "titulo": {
                        "description": "Título do livro",
                        "type": "string"
                    },
                    "autor_id": {
                        "description": "ID do autor do livro",
                        "type": "integer"
                    },
                    "categoria_id": {
                        "description": "ID da categoria do livro",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login com email e senha para obter o token de acesso",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}